[Bryce] Finish rigging robot model and improve dashboard performance

+++ src/main/java/frc/robot/Robot.java
@@ -83,7 +83,7 @@ public class Robot extends TimedRobot {
     if (netCycle == 0) {
       netTable.getEntry("alliance").setInteger(DriverStation.getAlliance().ordinal());
       netTable.getEntry("voltage").setDouble(RobotController.getBatteryVoltage());
-      netTable.getEntry("psi").setInteger((int) RobotContainer.compressor.getPressure());
+      netTable.getEntry("psi").setDouble(RobotContainer.compressor.getPressure());
       netTable.getEntry("time").setInteger((int) DriverStation.getMatchTime());
     }
   }
